refactor(테스트): 테스트 폴더 내 Domain 폴더 생성
