+++ __tests__/DomainTest/GiftTest.js
@@ -1,4 +1,4 @@
-import Gift from '../src/Domain/Benefit/Gift';
+import Gift from '../../src/Domain/Benefit/Gift';
 
 describe('증정 메뉴 정보 전달 테스트', () => {
   test('120,000원 주문 시 증정 메뉴는 샴페인 1개이다', () => {
